drm3d: remove unused blendv() and clear() code
Only gltex can use the drm3d interface, and gltex don't use
display_clear() nor display_blendv(), so they can safely be removed.
Also remove the corresponding shaders.
