+++ src/video/meson.build
@@ -42,10 +42,7 @@ video_deps = declare_dependency(
 if enable_video_drm3d
   mod_drm3d = shared_module('mod-drm3d', [
     'drm3d_video.c',
-    'drm3d_render.c',
     'mod_drm3d.c',
-    embed_gen.process('drm3d_blend.vert', extra_args: shader_regex),
-    embed_gen.process('drm3d_blend.frag', extra_args: shader_regex),
     ],
     name_prefix: '',
     dependencies: [video_deps, shl_deps, shl_gl_deps, libdrm_deps, egl_deps, gbm_deps, glesv2_deps],
